requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
diffusers==0.35.1
transformers==4.56.1
accelerate==1.10.1
peft==0.17.1
open-clip-torch==3.1.0
pillow==11.3.0
safetensors==0.6.2
numpy==2.2.6
tqdm==4.67.1
